feat: Added this workflow (also openapi wasnt up to date btw xddd)
